feat(web): extract hooks stats helpers into hooks-stats-lib
Move hook distribution and report-model assembly into hooks-stats-lib with expanded Vitest coverage.

Preserve the public @/lib/hooks-stats import surface.

Closes #4299
